+++ CHANGES.md
@@ -57,6 +57,8 @@
   * Add missing instances of SatModel for Char and String. Thanks to eax- on github
     for the contribution.
 
+  * Add support for symbolic comparison of regular expressions.
+
   * Export svToSV from Data.SBV.Dynamic. Thanks to Matt Parker for the PR.
 
 ### Version 8.16, 2021-08-18